Added test in stop to see if the pid even exists

+++ lib/cap/uni/tasks/unicorn.rake
@@ -16,7 +16,9 @@ namespace :unicorn do
   desc "Kill Unicorn server"
   task :stop do
     on roles(:all) do
-     execute "kill $(cat #{fetch(:unicorn_pid)})"
+     if test "[ -S fetch(:unicorn_pid) ]" then
+       execute "kill $(cat #{fetch(:unicorn_pid)})"
+     end
     end
   end
